feat : 매운맛 model 수정

+++ app/models/chat.py
@@ -15,7 +15,7 @@ class Chat(Base):
     updated_at = Column(DateTime, nullable=False, default=datetime.now, onupdate=datetime.now)
     name = Column(String(45), nullable=False)
     topic = Column(String(10), nullable=True)
-    spicy = Column(Integer, nullable=True)
+    spicy = Column(Float, nullable=True)
 
     character = relationship("Character", back_populates="chats")
     bubbles = relationship("Bubble", back_populates="chat")
